Adds tests for actual scicrunch service. Intended for manual exploratory testing of the actual service API
